Added options to save unassembled frames and histograms with sub-ADU precision, fixed a bug in cspadsubtractbehindwires

+++ source/libcheetah/include/detectorObject.h
@@ -272,7 +272,7 @@ public:
 	int		histogram;
 	long	histogramMin;
 	long	histogramNbins;
-	long	histogramBinSize;
+	float   histogramBinSize;
 	long	histogram_fs_min;
 	long	histogram_fs_max;
 	long	histogram_ss_min;
